changed integration test case class names.
