+++ integration-tests/src/test/resources/integration-test-suite.xml
@@ -21,8 +21,8 @@ limitations under the License.
     <test name="wso2as-test">
         <classes>
             <class name="org.wso2.appserver.test.integration.utils.ConfigurationLoaderTestCase"/>
-            <class name="org.wso2.appserver.test.integration.sample.SampleTestIT"/>
-            <class name="org.wso2.appserver.test.integration.loader.WebappLoaderIT"/>
+            <class name="org.wso2.appserver.test.integration.status.HTTPStatusTestCase"/>
+            <class name="org.wso2.appserver.test.integration.loader.WebappLoaderTestCase"/>
             <!--<class name="org.wso2.appserver.test.integration.statisticspublishing.StatisticsPublisherTestIT"/>-->
         </classes>
     </test>
